Fix next version number to 0.2.3
I saw 0.3.0 in the changelog and assumed that it was released into
kilted, but I think what happened was that the minor version number was
bumped on rolling and then kilted was branched off. So even though the
changelog shows there was a 0.3.0, that version was only released into
rolling, not into kilted. Checking `apt-cache show
ros-kilted-gz-cmake-vendor` shows that the last release was 0.2.2.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>